fix bug

+++ BmjAppBackend/database/migrations/2024_12_29_150459_create_detail_quotations.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->boolean('is_return')->nullable();
             $table->string('service')->nullable();
             $table->decimal('service_price', 15, 2)->nullable();
-            $table->integer('unit_price')->nullable();
+            $table->decimal('unit_price', 15, 2)->nullable();
             $table->timestamps();
         });
     }
